+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/informers/SharedInformerFactory.java
@@ -17,20 +17,28 @@ package io.fabric8.kubernetes.client.informers;
 
 import static io.fabric8.kubernetes.client.utils.KubernetesResourceUtil.inferListType;
 
+import io.fabric8.kubernetes.api.model.GenericKubernetesResource;
+import io.fabric8.kubernetes.api.model.GenericKubernetesResourceList;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.KubernetesResourceList;
 import io.fabric8.kubernetes.api.model.ListOptions;
 import io.fabric8.kubernetes.client.Config;
 import io.fabric8.kubernetes.client.ConfigBuilder;
 import io.fabric8.kubernetes.client.CustomResource;
+import io.fabric8.kubernetes.client.Handlers;
+import io.fabric8.kubernetes.client.ResourceHandler;
+import io.fabric8.kubernetes.client.SharedInformerOperationsImpl;
 import io.fabric8.kubernetes.client.Watch;
 import io.fabric8.kubernetes.client.Watcher;
+import io.fabric8.kubernetes.client.dsl.Resource;
 import io.fabric8.kubernetes.client.dsl.base.BaseOperation;
 import io.fabric8.kubernetes.client.dsl.base.CustomResourceDefinitionContext;
 import io.fabric8.kubernetes.client.dsl.base.OperationContext;
 import io.fabric8.kubernetes.client.informers.impl.DefaultSharedIndexInformer;
+import io.fabric8.kubernetes.client.utils.ApiVersionUtil;
 import io.fabric8.kubernetes.client.utils.Utils;
 import io.fabric8.kubernetes.internal.KubernetesDeserializer;
+
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -57,8 +65,6 @@ public class SharedInformerFactory extends BaseOperation {
 
   private final ExecutorService informerExecutor;
 
-  private final BaseOperation baseOperation;
-
   private final ConcurrentLinkedQueue<SharedInformerEventListener> eventListeners = new ConcurrentLinkedQueue<>();
 
   private boolean allowShutdown = true;
@@ -82,7 +88,6 @@ public class SharedInformerFactory extends BaseOperation {
     super(new OperationContext().withOkhttpClient(okHttpClient).withConfig(configuration));
     initOperationContext(configuration);
     this.informerExecutor = threadPool;
-    this.baseOperation = this.newInstance(context);
     this.namespace = null;
   }
 
@@ -122,7 +127,7 @@ public class SharedInformerFactory extends BaseOperation {
    * @return the shared index informer
    */
   public synchronized <T extends HasMetadata> SharedIndexInformer<T> sharedIndexInformerFor(Class<T> apiTypeClass, long resyncPeriodInMillis) {
-    return sharedIndexInformerFor(apiTypeClass, inferListType(apiTypeClass), null, resyncPeriodInMillis);
+    return sharedIndexInformerFor(apiTypeClass, inferListType(apiTypeClass), null, resyncPeriodInMillis, Utils.isResourceNamespaced(apiTypeClass));
   }
 
   /**
@@ -137,7 +142,7 @@ public class SharedInformerFactory extends BaseOperation {
    * @return the shared index informer
    */
   public synchronized <T extends HasMetadata> SharedIndexInformer<T> sharedIndexInformerFor(Class<T> apiTypeClass, OperationContext operationContext, long resyncPeriodInMillis) {
-    return sharedIndexInformerFor(apiTypeClass, inferListType(apiTypeClass), operationContext, resyncPeriodInMillis);
+    return sharedIndexInformerFor(apiTypeClass, inferListType(apiTypeClass), operationContext, resyncPeriodInMillis, Utils.isResourceNamespaced(apiTypeClass));
   }
 
   /**
@@ -165,7 +170,28 @@ public class SharedInformerFactory extends BaseOperation {
       .withApiGroupName(customResourceContext.getGroup())
       .withApiGroupVersion(customResourceContext.getVersion())
       .withPlural(customResourceContext.getPlural())
-      .withIsNamespaceConfiguredFromGlobalConfig(context.isNamespaceFromGlobalConfig()), resyncPeriodInMillis);
+      .withIsNamespaceConfiguredFromGlobalConfig(context.isNamespaceFromGlobalConfig()), resyncPeriodInMillis, Utils.isResourceNamespaced(apiTypeClass));
+  }
+
+  /**
+   * Constructs and returns a shared index informer with resync period specified for a Custom Resource. You
+   * can use it for scenarios where you don't have a POJO for your custom type by specifying group, version and plural in
+   * {@link CustomResourceDefinitionContext}
+   *
+   * <b>Note:</b>It watches for events in <b>ALL NAMESPACES</b>.
+   *
+   * @param genericResourceContext object containing details about resource like apiGroup, version and plural, etc.
+   * @param resyncPeriodInMillis resync period in milliseconds.
+   * @return {@link SharedIndexInformer} for GenericKubernetesResource
+   */
+  public synchronized SharedIndexInformer<GenericKubernetesResource> sharedIndexInformerForCustomResource(CustomResourceDefinitionContext genericResourceContext, long resyncPeriodInMillis) {
+    validateCustomResourceProvided(genericResourceContext);
+    return sharedIndexInformerFor(GenericKubernetesResource.class, GenericKubernetesResourceList.class, new OperationContext()
+      .withApiGroupName(genericResourceContext.getGroup())
+      .withApiGroupVersion(genericResourceContext.getVersion())
+      .withPlural(genericResourceContext.getPlural())
+      .withNamespace(this.namespace)
+      .withIsNamespaceConfiguredFromGlobalConfig(context.isNamespaceFromGlobalConfig()), resyncPeriodInMillis, genericResourceContext.isNamespaceScoped());
   }
 
   /**
@@ -178,7 +204,7 @@ public class SharedInformerFactory extends BaseOperation {
    */
   public synchronized <T extends CustomResource<?,?>> SharedIndexInformer<T> sharedIndexInformerForCustomResource(
     Class<T> apiTypeClass, OperationContext operationContext, long resyncPeriodInMillis) {
-    return sharedIndexInformerFor(apiTypeClass, inferListType(apiTypeClass), operationContext, resyncPeriodInMillis);
+    return sharedIndexInformerFor(apiTypeClass, inferListType(apiTypeClass), operationContext, resyncPeriodInMillis, Utils.isResourceNamespaced(apiTypeClass));
   }
 
   /**
@@ -208,7 +234,7 @@ public class SharedInformerFactory extends BaseOperation {
    * @return the shared index informer
    */
   public synchronized <T extends CustomResource<?,?>, L extends KubernetesResourceList<T>> SharedIndexInformer<T> sharedIndexInformerForCustomResource(Class<T> apiTypeClass, Class<L> apiListTypeClass, long resyncPeriodInMillis) {
-    return sharedIndexInformerFor(apiTypeClass, apiListTypeClass, null, resyncPeriodInMillis);
+    return sharedIndexInformerFor(apiTypeClass, apiListTypeClass, null, resyncPeriodInMillis, Utils.isResourceNamespaced(apiTypeClass));
   }
 
   /**
@@ -224,8 +250,8 @@ public class SharedInformerFactory extends BaseOperation {
    * @param <L> the type's list parameter (should extend {@link io.fabric8.kubernetes.api.model.KubernetesResourceList}
    * @return the shared index informer
    */
-  private synchronized <T extends HasMetadata, L extends KubernetesResourceList<T>> SharedIndexInformer<T> sharedIndexInformerFor(Class<T> apiTypeClass, Class<L> apiListTypeClass, OperationContext operationContext, long resyncPeriodInMillis) {
-    ListerWatcher<T, L> listerWatcher = listerWatcherFor(apiTypeClass, apiListTypeClass);
+  private synchronized <T extends HasMetadata, L extends KubernetesResourceList<T>> SharedIndexInformer<T> sharedIndexInformerFor(Class<T> apiTypeClass, Class<L> apiListTypeClass, OperationContext operationContext, long resyncPeriodInMillis, boolean isResourceNamespaced) {
+    ListerWatcher<T, L> listerWatcher = listerWatcherFor(apiTypeClass, apiListTypeClass, isResourceNamespaced);
     OperationContext context = this.context.withApiGroupName(HasMetadata.getGroup(apiTypeClass))
       .withApiGroupVersion(HasMetadata.getVersion(apiTypeClass))
       .withPlural(HasMetadata.getPlural(apiTypeClass))
@@ -248,19 +274,19 @@ public class SharedInformerFactory extends BaseOperation {
     return informer;
   }
 
-  private <T extends HasMetadata, L extends KubernetesResourceList<T>> ListerWatcher<T, L> listerWatcherFor(Class<T> apiTypeClass, Class<L> apiListTypeClass) {
+  private <T extends HasMetadata, L extends KubernetesResourceList<T>> ListerWatcher<T, L> listerWatcherFor(Class<T> apiTypeClass, Class<L> apiListTypeClass, boolean isResourceNamespaced) {
 
     return new ListerWatcher<T, L>() {
       @Override
       public L list(ListOptions params, String namespace, OperationContext context) {
-        BaseOperation<T, L, ?> listBaseOperation = getConfiguredBaseOperation(namespace, context, apiTypeClass, apiListTypeClass);
+        BaseOperation<T, L, ?> listBaseOperation = getConfiguredBaseOperation(namespace, context, apiTypeClass, apiListTypeClass, isResourceNamespaced);
         registerKindToKubernetesDeserializer(apiTypeClass);
         return listBaseOperation.list();
       }
 
       @Override
       public Watch watch(ListOptions params, String namespace, OperationContext context, Watcher<T> resourceWatcher) {
-        BaseOperation<T, L, ?> watchBaseOperation = getConfiguredBaseOperation(namespace, context, apiTypeClass, apiListTypeClass);
+        BaseOperation<T, L, ?> watchBaseOperation = getConfiguredBaseOperation(namespace, context, apiTypeClass, apiListTypeClass, isResourceNamespaced);
         registerKindToKubernetesDeserializer(apiTypeClass);
         return watchBaseOperation.watch(params.getResourceVersion(), resourceWatcher);
       }
@@ -372,23 +398,23 @@ public class SharedInformerFactory extends BaseOperation {
     return key.contains(plural);
   }
 
-    private <T extends HasMetadata, L extends KubernetesResourceList<T>> BaseOperation<T, L, ?> getConfiguredBaseOperation(String namespace, OperationContext context, Class<T> apiTypeClass, Class<L> apiListTypeClass) {
-    BaseOperation<T, L, ?> baseOperationWithContext;
+  private <T extends HasMetadata, L extends KubernetesResourceList<T>> BaseOperation<T, L, ?> getConfiguredBaseOperation(String namespace, OperationContext context, Class<T> apiTypeClass, Class<L> apiListTypeClass, boolean isNamespacedScoped) {
+    SharedInformerOperationsImpl<T, L, Resource<T>> sharedInformerOperations;
     // Avoid adding Namespace if it's picked from Global Configuration
     if (context.isNamespaceFromGlobalConfig()) {
       // SharedInformer default behavior is to watch in all namespaces
       // unless we specify namespace explicitly in OperationContext
-      baseOperationWithContext = baseOperation.newInstance(context
+      sharedInformerOperations = new SharedInformerOperationsImpl<>(context
         .withConfig(new ConfigBuilder(config)
           .withNamespace(null)
           .build())
-        .withNamespace(null));
+        .withNamespace(null), isNamespacedScoped);
     } else {
-      baseOperationWithContext = baseOperation.newInstance(context.withNamespace(namespace));
+      sharedInformerOperations = new SharedInformerOperationsImpl<>(context.withNamespace(namespace), isNamespacedScoped);
     }
-    baseOperationWithContext.setType(apiTypeClass);
-    baseOperationWithContext.setListType(apiListTypeClass);
-    return baseOperationWithContext;
+    sharedInformerOperations.setType(apiTypeClass);
+    sharedInformerOperations.setListType(apiListTypeClass);
+    return sharedInformerOperations;
   }
 
   private void initOperationContext(Config configuration) {
@@ -400,6 +426,17 @@ public class SharedInformerFactory extends BaseOperation {
   private <T extends HasMetadata> void registerKindToKubernetesDeserializer(Class<T> apiTypeClass) {
     if (CustomResource.class.isAssignableFrom(apiTypeClass)) {
       KubernetesDeserializer.registerCustomKind(HasMetadata.getApiVersion(apiTypeClass), apiTypeClass.getSimpleName(), apiTypeClass);
+    }
+  }
+
+  private void validateCustomResourceProvided(CustomResourceDefinitionContext genericResourceContext) {
+    if (Utils.isNullOrEmpty(genericResourceContext.getKind())) {
+      throw new IllegalArgumentException("CustomResourceDefinitionContext.kind: required value.");
+    }
+    ResourceHandler<HasMetadata, ?> resourceHandler = Handlers.get(genericResourceContext.getKind(), ApiVersionUtil.joinApiGroupAndVersion(genericResourceContext.getGroup(), genericResourceContext.getVersion()));
+
+    if (resourceHandler != null) {
+      throw new IllegalArgumentException("Using sharedIndexInformerDynamicResource for core type. Please use sharedIndexInformerFor(Class<T>, long) instead.");
     }
   }
 }