+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 * Add DSL support for `user.openshift.io/v1` Identity in OpenShiftClient DSL
 * Add DSL support for OpenShift Whereabouts CNI Model `whereabouts.cni.cncf.io` to OpenShiftClient DSL
 * Add DSL support for OpenShift Kube Storage Version Migrator resources in OpenShiftClient DSL
+* Fix #3228: Add support for Dynamic informers for custom resources in KubernetesClient
 
 #### _**Note**_: Breaking changes in the API
 ##### DSL Changes: